Only delete messages that aren't pinned

+++ bot-discord.js
@@ -97,7 +97,9 @@ async function postEventMsg() {
         await channel.fetchMessages();
         //Delete old messages
         for (let message of channel.messages.values()) {
-            await message.delete();
+            if (!message.pinned) {
+                await message.delete();
+            }
         }
         //Create new messages
         for (let i = 0; i < events.length; i++) {
